detection/requirements.txt
Flask
matplotlib>=2.2.3
networkx==2.5
numpy==1.19.5
opencv-contrib-python==4.5.1.48
opencv-python>=3.4.2.16
Pillow==8.1.0
pixellib==0.5.2
pytz==2020.5
PyWavelets==1.1.1
QtPy==1.9.0
requests-oauthlib==1.3.0
scikit-image==0.18.1
scikit-learn==0.24.1
scikit-network==0.20.0
Shapely==1.7.1
tensorflow>=2.4.1